requirements.txt
-i https://pypi.org/simple
absl-py==0.10.0
asgiref==3.2.10; python_version >= '3.5'
astunparse==1.6.3
cachetools==4.1.1; python_version ~= '3.5'
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
gast==0.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.1
google-auth==1.21.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
imutils==0.5.3
keras-preprocessing==1.1.2
kiwisolver==1.2.0; python_version >= '3.6'
markdown==3.2.2; python_version >= '3.5'
matplotlib==3.3.1
mysql-connector-python==8.0.21
mysql-connector-python-rf==2.2.2
numpy==1.18.5; python_version >= '3.5'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
opencv-python==4.4.0.42
opt-einsum==3.3.0; python_version >= '3.5'
pandas==1.1.1
pillow==7.2.0; python_version >= '3.5'
protobuf==3.13.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.1
requests-oauthlib==1.3.0
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.6; python_version >= '3.5'
scipy==1.4.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard-plugin-wit==1.7.0
tensorboard==2.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorflow-estimator==2.3.0
tensorflow==2.3.0
termcolor==1.1.0
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wheel==0.35.1; python_version >= '3'
wrapt==1.12.1
PyQt5~=5.15.1
dlib~=19.21.0
